requirements.txt
fastapi
jsmin
google-generativeai
python-dotenv